back/requirements.txt
django==3.1.6
gunicorn==20.*
django_environ==0.4.5
psycopg2-binary==2.8.4
six==1.13.0
djangorestframework_simplejwt==4.*
djangorestframework==3.*
django_extensions==3.*
requests==2.*
pytest==6.*
pytest-cov==2.*
pagarme-python==4.*
django-cors-headers==3.*
whitenoise
dj-static
